добавлена авторизация и регистрация

+++ controllers/cards.js
@@ -1,33 +1,39 @@
 const Card = require('../models/card');
-const { checkError } = require('../utils/functions');
+const NotFoundError = require('../errors/not-found-err');
+const ForbiddenError = require('../errors/forbidden-err');
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
 
   Card.create({ name, link, owner: req.user._id })
     .then((card) => res.send(card))
-    .catch((err) => res.status(checkError(err)).send({ message: res.statusCode === 500 ? 'Произошла ошибка' : err.message }));
+    .catch(next);
 };
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((card) => {
       res.send({ data: card });
     })
-    .catch((err) => res.status(checkError(err)).send({ message: res.statusCode === 500 ? 'Произошла ошибка' : err.message }));
+    .catch(next);
 };
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        const error = new Error('Not found');
-        error.name = 'NotFoundError';
+        const error = new NotFoundError('Not found');
         throw error;
       }
+      if (card.owner.toString() !== req.user._id) {
+        throw new ForbiddenError('Удалять можно только свою карточку');
+      }
+      Card.findByIdAndRemove(req.params.cardId);
+    })
+    .then((card) => {
       res.send({ data: card });
     })
-    .catch((err) => res.status(checkError(err)).send({ message: res.statusCode === 500 ? 'Произошла ошибка' : err.message }));
+    .catch(next);
 };
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -35,16 +41,15 @@ module.exports.likeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        const error = new Error('Not found');
-        error.name = 'NotFoundError';
+        const error = new NotFoundError('Not found');
         throw error;
       }
       res.send({ data: card });
     })
-    .catch((err) => res.status(checkError(err)).send({ message: res.statusCode === 500 ? 'Произошла ошибка' : err.message }));
+    .catch(next);
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -52,11 +57,10 @@ module.exports.dislikeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        const error = new Error('Not found');
-        error.name = 'NotFoundError';
+        const error = new NotFoundError('Not found');
         throw error;
       }
       res.send({ data: card });
     })
-    .catch((err) => res.status(checkError(err)).send({ message: res.statusCode === 500 ? 'Произошла ошибка' : err.message }));
+    .catch(next);
 };
